Created promise for build sourcelist and then chained in dashboard controller
