+++ services/dataretrieval/connectedSources.js
@@ -65,33 +65,3 @@ exports.refresh_connected_sources = function(req, res) {
 
   res.sendStatus(200);
 };
-
-// Build and format the users current source source list
-exports.build_sourceList = function(req, res) {
-  // Build the sourceList of users connected data sources
-  var humanId = req.user.humanapi.humanId;
-  console.log('ID for source list is: ' + humanId);
-  var sourceList = [];
-
-  function getSourceList(humanId) {
-    Sources.findOne({"humanId": humanId}, function(err, record) {
-      if (err) {
-        return console.log(err);
-      }
-      if (!record) {
-        sourceList.push('Please refresh connected data sources.');
-        res.status(200).send(sourceList);
-      }
-      else {
-        console.log('Sources object is: ' + record);
-        record.sources.forEach(function(source) {
-          var sourceDetail = source.source + ' has been connected since ' + source.connectedSince;
-          sourceList.push(sourceDetail);
-        });
-        console.log('Source list is: ' + sourceList);
-        res.status(200).send(sourceList);
-      }
-    });
-  };
-  getSourceList(humanId);
-};
